requirements.txt
# Python Version: 3.11
numpy<2
scipy==1.16.3
scikit-surprise==1.1.4
pandas==2.3.3
pathlib==1.0.1
matplotlib==3.10.8
seaborn==0.13.2
pickle-warehouse==0.1.1
ipython